+++ examples/nodejs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @klever/connect-examples-nodejs
 
+## 1.0.3
+
+### Patch Changes
+
+- @klever/connect@0.1.7
+
 ## 1.0.2
 
 ### Patch Changes